+++ main_program/Library/payment_framework/CHANGELOG.md
@@ -2,3 +2,5 @@
 ## [1.0.0] -2025-09-19
 ### Added
 - Added `pay_money` function to handle customer payments and update balance data, if payment successfully return True else return False
+### Changed
+- `pay_money` will not print any message,it will raise error